+++ src/assets/Objects/Genres.tsx
@@ -1,0 +1,14 @@
+export const Genres = [
+	'드라마',
+	'멜로/로맨스',
+	'범죄',
+	'스릴러',
+	'사극',
+	'성인/애로',
+	'애니메이션',
+	'액션',
+	'코미디',
+	'판타지',
+	'SF',
+	'기타',
+];
